Change size_t to unsigned int for name_length in operator overloads for Activity class

+++ Activity.cpp
@@ -16,7 +16,7 @@ std::string Activity::get_name() {
 
 
 std::ostream& operator<<(std::ostream &out, const Activity &activity) {
-	size_t name_length = activity.name.size();
+	unsigned int name_length = activity.name.size();
 	out.write(reinterpret_cast<char*>(&name_length), sizeof(name_length));
 	out.write(activity.name.c_str(), name_length);
 
@@ -26,7 +26,7 @@ std::ostream& operator<<(std::ostream &out, const Activity &activity) {
 }
 
 std::istream& operator>>(std::istream &in, Activity &activity) {
-	size_t name_length;
+	unsigned int name_length;
 	in.read(reinterpret_cast<char*>(&name_length), sizeof(name_length));
 	activity.name.resize(name_length);
 	in.read(&activity.name[0], name_length);
